% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{AddLooMafCol}
\alias{AddLooMafCol}
\title{add a leave-one-out minor allele freq. column with values to genotype table.}
\usage{
AddLooMafCol(geno, n.samples)
}
\arguments{
\item{geno}{data frame of genotypes.}

\item{n.samples}{total number of samples.}
}
\description{
add a leave-one-out minor allele freq. column with values to genotype table.
}

