% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{ApplyThreshold}
\alias{ApplyThreshold}
\title{subset on input table on a user defined condition.}
\usage{
ApplyThreshold(tbl, col.name, cmp, val)
}
\arguments{
\item{tbl}{input data frame.}

\item{col.name}{column name to subset input data frame on.}

\item{cmp}{string representing which operator to apply on df column.}

\item{val}{value linked to input tbl by cmp operator.}
}
\description{
subset on input table on a user defined condition.
}

