% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{BurdenTest}
\alias{BurdenTest}
\title{perform a case/control burden test on aggregated rare variant sequence data.}
\usage{
BurdenTest(mat, genes, cases, ctrls, collapse.thresh = 0, ...)
}
\arguments{
\item{mat}{collapsing matrix data structure.}

\item{genes}{list of genes to include in burden test.}

\item{cases}{vector of case names.}

\item{ctrls}{vector of ctrl names.}

\item{collapse.thresh}{integer threshold for inclusion of sample genotype,
if >0 then collapsing analyses are used, else 
poisson tests used (case vs. ctrl variant rate).}

\item{...}{additional args for fisher.test or poisson.test}
}
\description{
perform a case/control burden test on aggregated rare variant sequence data.
}

