% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsingmatrix.R
\name{CollapsingMatrixRead}
\alias{CollapsingMatrixRead}
\title{read in a collapsing matrix file from file (gzip detection supported) into
CollapsingMatrix data frame.}
\usage{
CollapsingMatrixRead(mat.filename)
}
\arguments{
\item{mat.filename}{name of collapsing matrix file.}
}
\description{
read in a collapsing matrix file from file (gzip detection supported) into
CollapsingMatrix data frame.
}

