% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsingmatrix.R
\name{CollapsingMatrixRecode}
\alias{CollapsingMatrixRecode}
\title{recode a collapsing matrix using a qualvar data frame.}
\usage{
CollapsingMatrixRecode(mat, qualvar)
}
\arguments{
\item{mat}{collapsing matrix data frame.}

\item{qualvar}{qualvar data frame.}
}
\description{
recode a collapsing matrix using a qualvar data frame.
}

