% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsingmatrix.R
\name{CollapsingMatrixShrink}
\alias{CollapsingMatrixShrink}
\title{take a subset of collapsing matrix where each gene has at least one 
sample with a qualifying variant and each sample has at least one 
qualifying variant in a gene.}
\usage{
CollapsingMatrixShrink(mat)
}
\arguments{
\item{mat}{collapsing matrix.}
}
\description{
take a subset of collapsing matrix where each gene has at least one 
sample with a qualifying variant and each sample has at least one 
qualifying variant in a gene.
}

