% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsingmatrix.R
\name{CollapsingMatrixWrite}
\alias{CollapsingMatrixWrite}
\title{write collapsing matrix to output filename (gzipping supported).}
\usage{
CollapsingMatrixWrite(mat, filename)
}
\arguments{
\item{mat}{collapsing matrix data frame.}

\item{filename}{output filename.}
}
\description{
write collapsing matrix to output filename (gzipping supported).
}

