% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{GetAlleleCounts}
\alias{GetAlleleCounts}
\title{for input variant calls and genotypes, return genotype counts, taking into
account the chromosome and sample gender for each variant call.   For 
example, a 'homozygous' on an x chromosome of a male would be an allele 
count of only 1, rather than 2, whereas a female homozygous call on an x
chromosome would be an allele count of 2.Such rules do not apply to 
autosomes, where each sample should have two copies of each.}
\usage{
GetAlleleCounts(sample.ID, variant.ID, genotype, sampleped)
}
\arguments{
\item{sample.ID}{vector of samples IDs}

\item{variant.ID}{vector of variant IDs}

\item{genotype}{vector of genotype strings (het / hom)}

\item{sampleped}{data frame of samples in sampleped format}
}
\description{
for input variant calls and genotypes, return genotype counts, taking into
account the chromosome and sample gender for each variant call.   For 
example, a 'homozygous' on an x chromosome of a male would be an allele 
count of only 1, rather than 2, whereas a female homozygous call on an x
chromosome would be an allele count of 2.Such rules do not apply to 
autosomes, where each sample should have two copies of each.
}

