% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkbed.R
\name{PlinkBedAlleleSum}
\alias{PlinkBedAlleleSum}
\title{get sum of allele counts across plinkbed for user-defined 
samples and variants.}
\usage{
PlinkBedAlleleSum(pb, samples = NULL, vars = NULL)
}
\arguments{
\item{pb}{plinkbed, a snpMatrix object courtesy of the snpStats pkg}

\item{samples}{vector of samples to subset on.}

\item{vars}{vector of variants to subset on.}
}
\description{
get sum of allele counts across plinkbed for user-defined 
samples and variants.
}

