% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkbed.R
\name{PlinkBedCalls}
\alias{PlinkBedCalls}
\title{return variant calls per sample that match user-defined genotype.}
\usage{
PlinkBedCalls(pb, samples, vars, gt = 1, as.df = F)
}
\arguments{
\item{pb}{plinkbed, a snpMatrix object courtesy of the snpStats pkg}

\item{samples}{a vector of samples to subset on.}

\item{vars}{a vector of variants to subset on.}

\item{gt}{genotype count to subset on.}

\item{as.df}{return calls in data.frame format?}
}
\description{
return variant calls per sample that match user-defined genotype.
}

