% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkbed.R
\name{PlinkBedGtsTransform}
\alias{PlinkBedGtsTransform}
\title{return genotypes from plinkbed into counts that are expected.}
\usage{
PlinkBedGtsTransform(gts)
}
\arguments{
\item{gts}{vector of genotypes from plinkbed snpMatrix obj.}
}
\description{
return genotypes from plinkbed into counts that are expected.
}

