% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkbed.R
\name{PlinkBedInit}
\alias{PlinkBedInit}
\title{functions for working with plinkbed, a snpMatrix object courtesy of the
snpStats pkg.}
\usage{
PlinkBedInit(bed.file, ...)
}
\arguments{
\item{bed.file}{plink bed filename.}

\item{...}{variables to be passed to the read.plink function in snpStats.}
}
\description{
initialize plinkbed.
}

