% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkbed.R
\name{PlinkBedSubset}
\alias{PlinkBedSubset}
\title{subset plinkbed on user-defined set of samples and variants.}
\usage{
PlinkBedSubset(pb, samples = NULL, vars = NULL)
}
\arguments{
\item{pb}{plinkbed, a snpMatrix object courtesy of the snpStats pkg}

\item{samples}{vector of samples to subset plinkbed on.}

\item{vars}{vector of variants to subset plinkbed on.}
}
\description{
subset plinkbed on user-defined set of samples and variants.
}

