% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plinkbed.R
\name{PlinkBedToQualvar}
\alias{PlinkBedToQualvar}
\title{convert plinkbed data to qualvar data frame and return.}
\usage{
PlinkBedToQualvar(pb, samples, varinfo, varinfo.class = NULL)
}
\arguments{
\item{pb}{plinkbed, a snpMatrix obj courtesy of the snpStats pkg.}

\item{samples}{a vector of samples to subset on.}

\item{varinfo}{a data frame in the varinfo format (varID,geneName,varClass).}

\item{varinfo.class}{variant class to subset on in varinfo data frame.}
}
\description{
convert plinkbed data to qualvar data frame and return.
}

