% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{ProcessString}
\alias{ProcessString}
\title{convert process input string into a form more agreeable with R.}
\usage{
ProcessString(str)
}
\arguments{
\item{str}{input string.}
}
\description{
convert process input string into a form more agreeable with R.
}

