% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualvar.R
\name{QualvarRead}
\alias{QualvarRead}
\title{read a file of format qualvar into a qualvar dataframe.}
\usage{
QualvarRead(qualvar.filename, samples = NULL, geneset = NULL)
}
\arguments{
\item{qualvar.filename}{name of qualvar file.}

\item{samples}{vector of samples to subset qualvar on.}

\item{geneset}{vector of genes to subset qualvar on.}
}
\description{
read a file of format qualvar into a qualvar dataframe.
}

