% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualvar.R
\name{QualvarSubset}
\alias{QualvarSubset}
\title{subset qualvar dataframe on user-defined sets of samples and genes.}
\usage{
QualvarSubset(qv, samples = NULL, geneset = NULL)
}
\arguments{
\item{qv}{qualvar data frame.}

\item{samples}{vector of samples to subset qualvar on.}

\item{geneset}{vector of genes to subset qualvar on.}
}
\description{
subset qualvar dataframe on user-defined sets of samples and genes.
}

