% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualvar.R
\name{QualvarToCollapsingMatrix}
\alias{QualvarToCollapsingMatrix}
\title{convert qualvar dataframe to collapsing matrix data frame.}
\usage{
QualvarToCollapsingMatrix(qv, samples, geneset)
}
\arguments{
\item{qv}{qualvar data frame.}

\item{samples}{vector of samples to subset qualvar on.}

\item{geneset}{vector of genes to subset qualvar on.}
}
\description{
convert qualvar dataframe to collapsing matrix data frame.
}

