% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{ReadConditionFile}
\alias{ReadConditionFile}
\title{read in a .condition file, where each line represents conditions to apply.}
\usage{
ReadConditionFile(condition.filename)
}
\arguments{
\item{condition.filename}{name of .condition file.}
}
\description{
read in a .condition file, where each line represents conditions to apply.
}

