% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{ReadGenesetsFile}
\alias{ReadGenesetsFile}
\title{read genesets file to a list of genesets.}
\usage{
ReadGenesetsFile(genesets.file)
}
\arguments{
\item{genesets.file}{name of .genesets file.}
}
\description{
read genesets file to a list of genesets.
}

