% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{ReadLargeTable}
\alias{ReadLargeTable}
\title{read a large table (can be gzipped and/or GB in size) into a data frame.}
\usage{
ReadLargeTable(filename, ...)
}
\arguments{
\item{filename}{name of table file.}

\item{...}{args to be passed to data.table fread function}
}
\description{
read a large table (can be gzipped and/or GB in size) into a data frame.
}

