% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{ReadListFile}
\alias{ReadListFile}
\title{read in a file where each line is a seperate entry into a vector.}
\usage{
ReadListFile(filename)
}
\arguments{
\item{filename}{name of input file.}
}
\description{
read in a file where each line is a seperate entry into a vector.
}

