% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{SamplepedIidFidList}
\alias{SamplepedIidFidList}
\title{table sampleped data frame and get list of individual ID -> family ID.}
\usage{
SamplepedIidFidList(sampleped)
}
\arguments{
\item{sampleped}{data frame of sample data in sampleped format.}
}
\description{
table sampleped data frame and get list of individual ID -> family ID.
}

