% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{StringToSet}
\alias{StringToSet}
\title{split string to vector}
\usage{
StringToSet(str, type = "character", delim = ",")
}
\arguments{
\item{str}{input string.}

\item{type}{type of variable that each output vector element should be.}

\item{delim}{delimiter character in input string.}
}
\description{
split string to vector
}

