% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{TblCndSubset}
\alias{TblCndSubset}
\title{subset on table based on rows in condition data frame.}
\usage{
TblCndSubset(tbl, cnd)
}
\arguments{
\item{tbl}{data frame of values with column names.}

\item{cnd}{data frame of conditions to subset tbl on.}
}
\description{
subset on table based on rows in condition data frame.
}

