% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{UpdateTable}
\alias{UpdateTable}
\title{update input table with values of res.list at row i}
\usage{
UpdateTable(tbl, row.i, res.list)
}
\arguments{
\item{tbl}{input table.}

\item{row.i}{row number in table to edit.}

\item{res.list}{list of results to iterate through and add vals to table.}
}
\description{
update input table with values of res.list at row i
}

