% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_functions.R
\name{WriteLargeTable}
\alias{WriteLargeTable}
\title{write large table to output file, with support for gzip.}
\usage{
WriteLargeTable(tbl, filename, ...)
}
\arguments{
\item{tbl}{data frame to be written to file.}

\item{filename}{name of output file to write to, ".gz" at end for gzip.}

\item{...}{parameters to be passed on to write.table function.}
}
\description{
write large table to output file, with support for gzip.
}

